Functioning camera WIP
