+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: plouda <marvin@42.fr>                      +#+  +:+       +#+         #
+#    By: plouda <plouda@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/03/04 15:40:17 by okraus            #+#    #+#              #
-#    Updated: 2023/12/08 11:15:33 by plouda           ###   ########.fr        #
+#    Updated: 2023/12/12 12:53:35 by plouda           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -71,7 +71,8 @@ SRC_S		=	minirt.c \
 				o_cylinder.c \
 				object_handler.c \
 				find_rays.c \
-				o_cylinder_disc.c
+				o_cylinder_disc.c \
+				camera.c
 
 # Formating
