+++ header/minirt.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minirt.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: plouda <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: plouda <plouda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/04 15:43:08 by okraus            #+#    #+#             */
-/*   Updated: 2023/12/11 09:05:02 by plouda           ###   ########.fr       */
+/*   Updated: 2023/12/12 12:54:46 by plouda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -244,5 +244,13 @@ void	define_topcap(t_cylinder *cylinder);
 void	get_discs(t_cylinder *cylinder);
 void	free_discs(t_cylinder *cylinder);
 
+// Camera
+double	**set_camera(t_camera *camera);
+t_vect3f	shift_origin(double **cam);
+void	change_ray_direction(double **cam, t_vect3f *direction, t_vect3f temp);
+void	shift_camera(t_master *master, mlx_key_data_t keydata);
+void	rotate_camera(t_master *master, mlx_key_data_t keydata);
+
+t_vect3f	cross_product(t_vect3f vect1, t_vect3f vect2);
 
 #endif
