+++ maps/cylinder2.rt
@@ -1,6 +1,6 @@
 A 0000.9999 55,5,70
 
-C  8000,0,99.00005 0.004,0.2,-1.0 150
+C  20,1,10 0,0,1 135
 L -4000,0,30 .17 255,255,255
 
 cy -56.07,22.38,-108.69 0.801,-0.534,0.267 20 7 10,0,255
