+++ src/find_rays.c
@@ -6,7 +6,7 @@
 /*   By: plouda <plouda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/04 09:47:21 by plouda            #+#    #+#             */
-/*   Updated: 2023/12/11 16:41:59 by plouda           ###   ########.fr       */
+/*   Updated: 2023/12/12 12:54:32 by plouda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,6 +65,16 @@ double	dot_product(t_vect3f vect1, t_vect3f vect2)
 	return (res);
 }
 
+t_vect3f	cross_product(t_vect3f vect1, t_vect3f vect2)
+{
+	t_vect3f	res;
+
+	res.x = vect1.y * vect2.z - vect1.z * vect2.y;
+	res.y = vect1.z * vect2.x - vect1.x * vect2.z;
+	res.z = vect1.x * vect2.y - vect1.y * vect2.x;
+	return (res);
+}
+
 t_vect3f	subtract_center(t_vect3f vect1, double *coords)
 {
 	t_vect3f	res;
@@ -72,16 +82,6 @@ t_vect3f	subtract_center(t_vect3f vect1, double *coords)
 	res.x = vect1.x - coords[X];
 	res.y = vect1.y - coords[Y];
 	res.z = vect1.z - coords[Z];
-	return (res);
-}
-
-t_vect3f	subtract_center_abs(t_vect3f vect1, double *coords)
-{
-	t_vect3f	res;
-
-	res.x = absf(vect1.x - coords[X]);
-	res.y = absf(vect1.y - coords[Y]);
-	res.z = absf(vect1.z - coords[Z]);
 	return (res);
 }
 
@@ -105,20 +105,14 @@ int	test_sphere(t_ray ray, t_sphere *sphere, double *t)
 	double	discr;
 	double	q;
 	t_vect3f	centered;
-	//t_vect3f	centered_abs;
 
-	//res = malloc(sizeof(int) * 2);
 	centered = subtract_center(ray.origin, sphere->coords);
-	//centered_abs = subtract_center_abs(ray->origin, sphere->coords);
 	a = dot_product(ray.direction, ray.direction);
 	b = 2 * dot_product(ray.direction, centered);
 	c = dot_product(centered, centered) - pow(sphere->diameter / 2, 2);
 	discr = pow(b, 2) - 4 * a * c;
 	if (discr < 0)
-	{
-		//free(res);
 		return (0);
-	}
 	else if (discr == 0)
 	{
 		res[0] = - 0.5 * b / a;
@@ -140,10 +134,7 @@ int	test_sphere(t_ray ray, t_sphere *sphere, double *t)
 		{
 			res[0] = res[1];
 			if (res[0] < 0)
-			{
-				//free(res);
 				return (0);
-			}
 		}
 		*t = res[0];
 		return (1);
@@ -290,26 +281,17 @@ int	test_cylinder(t_ray ray, t_cylinder *cylinder, double *t)
 	double	q;
 	t_vect3f	centered;
 	t_vect3f	axis;
-	//t_vect3f	point;
-	//double		max;
 
-	/* axis.x = cylinder->nvect[X];
-	axis.y = cylinder->nvect[Y];
-	axis.z = cylinder->nvect[Z]; */
 	axis.x = cylinder->normal->x;
 	axis.y = cylinder->normal->y;
 	axis.z = cylinder->normal->z;
-	//res = malloc(sizeof(int) * 2);
 	centered = subtract_center(ray.origin, cylinder->coords);
 	a = dot_product(ray.direction, ray.direction) - pow(dot_product(ray.direction, axis), 2);
 	b = 2 * (dot_product(ray.direction, centered) - (dot_product(ray.direction, axis) * dot_product(centered, axis)));
 	c = dot_product(centered, centered) - pow(dot_product(centered, axis), 2) - pow(cylinder->diameter / 2, 2);
 	discr = pow(b, 2) - 4 * a * c;
 	if (discr < 0)
-	{
-		//free(res);
 		return (0);
-	}
 	else if (discr == 0)
 	{
 		res[0] = - 0.5 * b / a;
@@ -336,10 +318,7 @@ int	test_cylinder(t_ray ray, t_cylinder *cylinder, double *t)
 		{
 			res[0] = res[1];
 			if (res[0] < 0)
-			{
-				//free(res);
 				return (0);
-			}
 		}
 		if (is_between_caps(cylinder->botcap, cylinder->topcap, ray, res[0]))
 		{
@@ -391,86 +370,6 @@ void	get_nearest_object(int flag, void *object_ptr, uint32_t *clr)
 	}
 }
 
-t_vect3f	cross_product(t_vect3f vect1, t_vect3f vect2)
-{
-	t_vect3f	res;
-
-	res.x = vect1.y * vect2.z - vect1.z * vect2.y;
-	res.y = vect1.z * vect2.x - vect1.x * vect2.z;
-	res.z = vect1.x * vect2.y - vect1.y * vect2.x;
-	return (res);
-}
-
-double	**set_camera(t_camera *camera)
-{
-	t_vect3f	tmp;
-	//t_vect3f	from;
-	//t_vect3f	to;
-	t_vect3f	forward;
-	t_vect3f	right;
-	t_vect3f	up;
-	double		**matrix;
-
-	matrix = malloc(sizeof(double *) * 4);
-	matrix[0] = malloc(sizeof(double) * 4);
-	matrix[1] = malloc(sizeof(double) * 4);
-	matrix[2] = malloc(sizeof(double) * 4);
-	matrix[3] = malloc(sizeof(double) * 4);
-	tmp.x = 0;
-	tmp.y = 1;
-	tmp.z = 0;
-	forward.x = camera->normal->x;
-	forward.y = camera->normal->y;
-	forward.z = camera->normal->z;
-	right = cross_product(tmp, forward);
-	normalize(&right, right.x, right.y, right.z);
-	up = cross_product(forward, right);
-	normalize(&up, up.x, up.y, up.z);
-	matrix[0][0] = right.x;
-	matrix[0][1] = right.y;
-	matrix[0][2] = right.z;
-	matrix[0][3] = 0;
-	matrix[1][0] = up.x;
-	matrix[1][1] = up.y;
-	matrix[1][2] = up.z;
-	matrix[1][3] = 0;
-	matrix[2][0] = forward.x;
-	matrix[2][1] = forward.y;
-	matrix[2][2] = forward.z;
-	matrix[2][3] = 0;
-	matrix[3][0] = camera->coords[X];
-	matrix[3][1] = camera->coords[Y];
-	matrix[3][2] = camera->coords[Z];
-	matrix[3][3] = 1;
-	/* printf("WorldToCamera:\n%.5f %.5f %.5f;\n%.5f %.5f %.5f;\n%.5f %.5f %.5f;\n%.5f %.5f %.5f;\n", \
-	matrix[0][0], matrix[0][1], matrix[0][2], matrix[1][0], matrix[1][1], matrix[1][2], \
-	matrix[2][0], matrix[2][1], matrix[2][2], matrix[3][0], matrix[3][1], matrix[3][2]); */
-	return (matrix);
-}
-
-/* since origin is 0,0,0, it's just shifting it to "from" point. Proper 
-matrix multiplication becomes relevant when:
-1/ origin is not 0,0,0
-2/ homogenous coordinate is not 1 (e.g. after projection matrix)
-*/
-t_vect3f	shift_origin(double **cam)
-{
-	t_vect3f	origin;
-
-	origin.x = cam[3][0];
-	origin.y = cam[3][1];
-	origin.z = cam[3][2];
-	return (origin);
-}
-
-void	change_ray_direction(double **cam, t_vect3f *direction, t_vect3f temp)
-{
-	direction->x = temp.x * cam[0][0] + temp.y * cam[1][0] + temp.z * cam[2][0];
-	direction->y = temp.x * cam[0][1] + temp.y * cam[1][1] + temp.z * cam[2][1];
-	direction->z = temp.x * cam[0][2] + temp.y * cam[1][2] + temp.z * cam[2][2];
-	normalize(direction, direction->x, direction->y, direction->z);
-}
-
 void	find_rays(t_master *master)
 {
 	int		i;
@@ -505,8 +404,8 @@ void	find_rays(t_master *master)
 	y = 0;
 	fov = master->rt->camera->fov;
 	ratio = (double)WIDTH / (double)HEIGHT; // assuming WIDTH > HEIGHT
-	camera_matrix = set_camera(master->rt->camera);
-	origin = shift_origin(camera_matrix);
+	camera_matrix = set_camera(master->rt->camera); // create camera matrix
+	origin = shift_origin(camera_matrix); // translate camera position
 	while (x < WIDTH)
 	{
 		y = 0;
@@ -521,14 +420,13 @@ void	find_rays(t_master *master)
 			rays[x][y].origin.y = origin.y;
 			rays[x][y].origin.z = origin.z;
 			// why cos(rad)? see https://www.permadi.com/tutorial/raycast/rayc8.html, alternatively change z to a lower value (e.g. -10)
-			px = (2. * ((x + 0.5) / (double)WIDTH) - 1.) * ratio * tan((rad(fov * 0.5)));
-			py = (1. - 2. * ((y + 0.5) / (double)HEIGHT)) * tan((rad(fov * 0.5)));
-			pz = -10; // change to calibrate focal length, should be -1
+			px = (2. * ((x + 0.5) / (double)WIDTH) - 1.) * ratio * tan(cos(rad(fov * 0.5)));
+			py = (1. - 2. * ((y + 0.5) / (double)HEIGHT)) * tan(cos(rad(fov * 0.5)));
+			pz = -1; // change to calibrate focal length, should be -1
 			direction.x = px;
 			direction.y = py;
 			direction.z = pz;
-			change_ray_direction(camera_matrix, &rays[x][y].direction, direction);
-			//normalize(&rays[x][y].direction, px, py, pz);
+			change_ray_direction(camera_matrix, &rays[x][y].direction, direction); // change direction based on camera position and its direction
 			i = 0;
 			while (i < master->rt->n_spheres)
 			{
