+++ src/minirt.c
@@ -6,7 +6,7 @@
 /*   By: plouda <plouda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/22 13:40:49 by okraus            #+#    #+#             */
-/*   Updated: 2023/12/11 18:14:51 by plouda           ###   ########.fr       */
+/*   Updated: 2023/12/12 12:49:54 by plouda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,38 +58,6 @@ int	load_file(char *file, t_rt *rt)
 	if (!flag)
 		load_data(rt, file, &flag);
 	return (flag);
-}
-
-void	shift_camera(t_master *master, mlx_key_data_t keydata)
-{
-	if (keydata.key == MLX_KEY_RIGHT)
-		master->rt->camera->coords[X] += 5;
-	if (keydata.key == MLX_KEY_LEFT)
-		master->rt->camera->coords[X] -= 5;
-	if (keydata.key == MLX_KEY_UP)
-		master->rt->camera->coords[Y] += 5;
-	if (keydata.key == MLX_KEY_DOWN)
-		master->rt->camera->coords[Y] -= 5;
-	if (keydata.key == MLX_KEY_KP_ADD)
-		master->rt->camera->coords[Z] -= 5;
-	if (keydata.key == MLX_KEY_KP_SUBTRACT)
-		master->rt->camera->coords[Z] += 5;
-	find_rays(master);
-}
-
-void	rotate_camera(t_master *master, mlx_key_data_t keydata)
-{
-	if (keydata.key == MLX_KEY_Q)
-		master->rt->camera->normal->x += 0.05;
-	if (keydata.key == MLX_KEY_E)
-		master->rt->camera->normal->x -= 0.05;
-	if (keydata.key == MLX_KEY_W)
-		master->rt->camera->normal->y -= 0.05;
-	if (keydata.key == MLX_KEY_S)
-		master->rt->camera->normal->y += 0.05;
-	normalize(master->rt->camera->normal, master->rt->camera->normal->x, \
-	master->rt->camera->normal->y, master->rt->camera->normal->z);
-	find_rays(master);
 }
 
 void	keyhook(mlx_key_data_t keydata, void *param)
